requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
devolo-home-control-api==0.17.4
idna==3.3
ifaddr==0.1.7
python-dotenv==0.19.2
requests==2.26.0
urllib3==1.26.7
websocket-client==1.2.1
zeroconf==0.37.0
